Remettre l'espacement de la grille dans le contexte de la grille
